move insertDocument from hook to util returning a Promise
